Move DataContainer to collections crate
Moving DataContainer to collections makes implementation of custom
store views much simpler since heay lifting around data is handled
there. Now you need to only know how to feed the data into container and
interprete changeset results
