+++ relm4-store-collections/Cargo.toml
@@ -1,6 +1,5 @@
 [package]
 name = "relm4-store-collections"
-publish = false
 description = "Custom collections required by the relm4-store"
 version = "0.1.0-beta.1"
 edition = "2021"
@@ -12,3 +11,8 @@ readme = "README.md"
 license = "Apache-2.0 OR MIT"
 
 [dependencies]
+reexport = { path = "../relm4-store-reexport", package = "relm4-store-reexport", version="0.1.0-beta.1" }
+record = { path = "../relm4-store-record", package = "relm4-store-record", version="0.1.0-beta.1" }
+
+[dev-dependencies]
+backend_dummy = { path = "../relm4-store-backend-dummy", package = "relm4-store-backend-dummy" }
