+++ relm4-store-collections/src/data_container/tests/insert_left.rs
@@ -1,37 +1,39 @@
-//! Contains tests for `[DataContainer::insert_right]`
+//! Contains tests for `[DataContainer::insert_left]`
 
 use backend_dummy::test_cases::TestRecord;
 use record::Record;
 
-use crate::WindowChangeset;
-use crate::implementation::data_container::DataContainer;
+use crate::data_container::WindowChangeset;
+use crate::data_container::DataContainer;
+
 use super::test_data::TestData;
 
 #[test]
-fn insert_right_first_record() {
+fn insert_left_first_record_at_0() {
     let mut dc: DataContainer<TestRecord> = DataContainer::new(10);
     let mut changeset: WindowChangeset<TestRecord> = WindowChangeset::default();
     let record: TestRecord = TestRecord::constant("First record to add");
     let records: Vec<TestRecord> = vec![record.clone()];
-    let position: usize = 0;
+    let position: usize = 1;
 
-    dc.insert_right(&mut changeset, position, records);
+    dc.insert_left(&mut changeset, position, records);
 
     dc.invariants();
+    assert_eq!(dc.len(), 1);
     assert_eq!(dc.data[&record.get_id()], record);
     assert_eq!(dc.order[0], record.get_id());
 }
 
 #[test]
-fn insert_right_second_two_records() {
+fn insert_left_second_two_records() {
     let mut dc: DataContainer<TestRecord> = DataContainer::new(10);
     let mut changeset: WindowChangeset<TestRecord> = WindowChangeset::default();
     let record1: TestRecord = TestRecord::constant("First record to add");
     let record2: TestRecord = TestRecord::constant("Second record to add");
     let records: Vec<TestRecord> = vec![record1.clone(), record2.clone()];
-    let position: usize = 0;
+    let position: usize = 2;
 
-    dc.insert_right(&mut changeset, position, records);
+    dc.insert_left(&mut changeset, position, records);
 
     dc.invariants();
     assert_eq!(dc.len(), 2);
@@ -42,16 +44,16 @@ fn insert_right_second_two_records() {
 }
 
 #[test]
-fn insert_right_three_records() {
+fn insert_left_three_records() {
     let mut dc: DataContainer<TestRecord> = DataContainer::new(10);
     let mut changeset: WindowChangeset<TestRecord> = WindowChangeset::default();
     let record1: TestRecord = TestRecord::constant("First record to add");
     let record2: TestRecord = TestRecord::constant("Second record to add");
     let record3: TestRecord = TestRecord::constant("Third record to add");
     let records: Vec<TestRecord> = vec![record1.clone(), record2.clone(), record3.clone()];
-    let position: usize = 0;
+    let position: usize = 3;
 
-    dc.insert_right(&mut changeset, position, records);
+    dc.insert_left(&mut changeset, position, records);
 
     dc.invariants();
     assert_eq!(dc.len(), 3);
@@ -86,13 +88,13 @@ mod max_size_3 {
         use super::MAX_SIZE;
 
         #[test]
-        fn add_at_0_elements_0() {
+        fn add_at_max_elements_0() {
             let TestData{ records, mut container } = TestData::new(RECORDS_CNT, MAX_SIZE);
             let mut changeset: WindowChangeset<TestRecord> = WindowChangeset::default();
             let new_records = vec![];
-            let position = 0;
+            let position = records.len()-1;
 
-            container.insert_right(&mut changeset, position, new_records);
+            container.insert_left(&mut changeset, position, new_records);
 
             assert_eq!(container.len(), 3);
 
@@ -106,76 +108,77 @@ mod max_size_3 {
         }
 
 
-        /// `|1, 2, 3| -> |4, 1, 2|`
+        /// `|1, 2, 3| -> |2, 3, 4|`
         #[test]
-        fn add_at_0_elements_1() {
+        fn add_at_max_elements_1() {
             let TestData{ records, mut container } = TestData::new(RECORDS_CNT, MAX_SIZE);
             let mut changeset: WindowChangeset<TestRecord> = WindowChangeset::default();
             let new_record = TestRecord::since("Record added at pos 0", 1);
             let new_records = vec![new_record.clone()];
-            let position = 0;
+            let position = records.len();
 
-            container.insert_right(&mut changeset, position, new_records);
+            container.insert_left(&mut changeset, position, new_records);
 
             assert_eq!(container.len(), 3);
 
             assert_eq!(container.data[&new_record.get_id()], new_record);
-            assert_eq!(container.data[&records[0].get_id()], records[0]);
             assert_eq!(container.data[&records[1].get_id()], records[1]);
+            assert_eq!(container.data[&records[2].get_id()], records[2]);
 
-            assert_eq!(container.order[0], new_record.get_id());
-            assert_eq!(container.order[1], records[0].get_id());
-            assert_eq!(container.order[2], records[1].get_id());
+            assert_eq!(container.order[0], records[1].get_id());
+            assert_eq!(container.order[1], records[2].get_id());
+            assert_eq!(container.order[2], new_record.get_id());
         }
 
-        /// `|1, 2, 3| -> |4, 5, 1|
+        /// `|1, 2, 3| -> |3, 4, 5|
         #[test]
-        fn add_at_0_elements_2() {
+        fn add_at_max_elements_2() {
             let TestData{ records, mut container } = TestData::new(RECORDS_CNT, MAX_SIZE);
             let mut changeset: WindowChangeset<TestRecord> = WindowChangeset::default();
             let new_record1 = TestRecord::since("Record added at pos 0 - 0", 1);
             let new_record2 = TestRecord::since("Record added at pos 0 - 1", 1);
             let new_records = vec![new_record1.clone(), new_record2.clone()];
-            let position = 0;
+            let position = records.len();
 
-            container.insert_right(&mut changeset, position, new_records);
-
+            
+            container.insert_left(&mut changeset, position, new_records);
+            
             assert_eq!(container.len(), 3);
-
+            
             assert_eq!(container.data[&new_record1.get_id()], new_record1);
             assert_eq!(container.data[&new_record2.get_id()], new_record2);
-            assert_eq!(container.data[&records[0].get_id()], records[0]);
-
-            assert_eq!(container.order[0], new_record1.get_id());
-            assert_eq!(container.order[1], new_record2.get_id());
-            assert_eq!(container.order[2], records[0].get_id());
+            assert_eq!(container.data[&records[2].get_id()], records[2]);
+            
+            assert_eq!(container.order[0], records[2].get_id());
+            assert_eq!(container.order[1], new_record1.get_id());
+            assert_eq!(container.order[2], new_record2.get_id());
         }
-
+        
         #[test]
-        fn add_at_0_elements_3() {
-            let TestData{ mut container, .. } = TestData::new(RECORDS_CNT, MAX_SIZE);
+        fn add_at_max_elements_3() {
+            let TestData{ mut container, records } = TestData::new(RECORDS_CNT, MAX_SIZE);
             let mut changeset: WindowChangeset<TestRecord> = WindowChangeset::default();
             let new_record1 = TestRecord::since("Record added at pos 0 - 0", 1);
             let new_record2 = TestRecord::since("Record added at pos 0 - 1", 1);
             let new_record3 = TestRecord::since("Record added at pos 0 - 2", 1);
             let new_records = vec![new_record1.clone(), new_record2.clone(), new_record3.clone()];
-            let position = 0;
-
-            container.insert_right(&mut changeset, position, new_records);
-
+            let position = records.len();
+            
+            container.insert_left(&mut changeset, position, new_records);
+            
             assert_eq!(container.len(), 3);
-
+            
             assert_eq!(container.data[&new_record1.get_id()], new_record1);
             assert_eq!(container.data[&new_record2.get_id()], new_record2);
             assert_eq!(container.data[&new_record3.get_id()], new_record3);
-
+            
             assert_eq!(container.order[0], new_record1.get_id());
             assert_eq!(container.order[1], new_record2.get_id());
             assert_eq!(container.order[2], new_record3.get_id());
         }
-
+        
         #[test]
-        fn add_at_0_elements_4() {
+        fn add_at_max_elements_4() {
             let TestData{ mut container, .. } = TestData::new(RECORDS_CNT, MAX_SIZE);
             let mut changeset: WindowChangeset<TestRecord> = WindowChangeset::default();
             let new_record1 = TestRecord::since("Record added at pos 0 - 0", 1);
@@ -188,24 +191,24 @@ mod max_size_3 {
                 new_record3.clone(), 
                 new_record4.clone()
             ];
-            let position = 0;
+            let position = RECORDS_CNT;
 
-            container.insert_right(&mut changeset, position, new_records);
-
+            container.insert_left(&mut changeset, position, new_records);
+            
             assert_eq!(container.len(), 3);
-
-            assert_eq!(container.data[&new_record1.get_id()], new_record1);
+            
             assert_eq!(container.data[&new_record2.get_id()], new_record2);
             assert_eq!(container.data[&new_record3.get_id()], new_record3);
-
-            assert_eq!(container.order[0], new_record1.get_id());
-            assert_eq!(container.order[1], new_record2.get_id());
-            assert_eq!(container.order[2], new_record3.get_id());
+            assert_eq!(container.data[&new_record4.get_id()], new_record4);
+            
+            assert_eq!(container.order[0], new_record2.get_id());
+            assert_eq!(container.order[1], new_record3.get_id());
+            assert_eq!(container.order[2], new_record4.get_id());
         }
 
         #[test]
-        fn add_at_0_elements_5() {
-            let TestData{ mut container, .. } = TestData::new(RECORDS_CNT, MAX_SIZE);
+        fn add_at_max_elements_5() {
+            let TestData{ mut container, ..} = TestData::new(RECORDS_CNT, MAX_SIZE);
             let mut changeset: WindowChangeset<TestRecord> = WindowChangeset::default();
             let new_record1 = TestRecord::since("Record added at pos 0 - 0", 1);
             let new_record2 = TestRecord::since("Record added at pos 0 - 1", 1);
@@ -219,24 +222,24 @@ mod max_size_3 {
                 new_record4.clone(),
                 new_record5.clone(),
             ];
-            let position = 0;
+            let position = RECORDS_CNT;
 
-            container.insert_right(&mut changeset, position, new_records);
+            container.insert_left(&mut changeset, position, new_records);
 
             assert_eq!(container.len(), 3);
 
-            assert_eq!(container.data[&new_record1.get_id()], new_record1);
-            assert_eq!(container.data[&new_record2.get_id()], new_record2);
             assert_eq!(container.data[&new_record3.get_id()], new_record3);
+            assert_eq!(container.data[&new_record4.get_id()], new_record4);
+            assert_eq!(container.data[&new_record5.get_id()], new_record5);
 
-            assert_eq!(container.order[0], new_record1.get_id());
-            assert_eq!(container.order[1], new_record2.get_id());
-            assert_eq!(container.order[2], new_record3.get_id());
+            assert_eq!(container.order[0], new_record3.get_id());
+            assert_eq!(container.order[1], new_record4.get_id());
+            assert_eq!(container.order[2], new_record5.get_id());
         }
     }
 
     #[test]
-    fn add_at_1() {
+    fn add_at_0() {
         
         let TestData{ records, mut container } = TestData::new(RECORDS_CNT, MAX_SIZE);
         let mut changeset: WindowChangeset<TestRecord> = WindowChangeset::default();
@@ -244,37 +247,37 @@ mod max_size_3 {
         let new_records = vec![new_record.clone()];
         let position = 1;
 
-        container.insert_right(&mut changeset, position, new_records);
+        container.insert_left(&mut changeset, position, new_records);
 
         assert_eq!(container.len(), 3);
 
         assert_eq!(container.data[&new_record.get_id()], new_record);
-        assert_eq!(container.data[&records[0].get_id()], records[0]);
         assert_eq!(container.data[&records[1].get_id()], records[1]);
+        assert_eq!(container.data[&records[2].get_id()], records[2]);
 
-        assert_eq!(container.order[0], records[0].get_id());
-        assert_eq!(container.order[1], new_record.get_id());
-        assert_eq!(container.order[2], records[1].get_id());
+        assert_eq!(container.order[0], new_record.get_id());
+        assert_eq!(container.order[1], records[1].get_id());
+        assert_eq!(container.order[2], records[2].get_id());
     }
 
     #[test]
-    fn add_at_2() {   
+    fn add_at_1() {   
         let TestData{ records, mut container } = TestData::new(RECORDS_CNT, MAX_SIZE);
         let mut changeset: WindowChangeset<TestRecord> = WindowChangeset::default();
-        let new_record = TestRecord::since("Record added at pos 0", 1);
+        let new_record = TestRecord::since("Record added at pos 1", 1);
         let new_records = vec![new_record.clone()];
         let position = 2;
 
-        container.insert_right(&mut changeset, position, new_records);
+        container.insert_left(&mut changeset, position, new_records);
 
         assert_eq!(container.len(), 3);
 
         assert_eq!(container.data[&new_record.get_id()], new_record);
-        assert_eq!(container.data[&records[0].get_id()], records[0]);
         assert_eq!(container.data[&records[1].get_id()], records[1]);
+        assert_eq!(container.data[&records[2].get_id()], records[2]);
 
-        assert_eq!(container.order[0], records[0].get_id());
-        assert_eq!(container.order[1], records[1].get_id());
-        assert_eq!(container.order[2], new_record.get_id());
+        assert_eq!(container.order[0], records[1].get_id());
+        assert_eq!(container.order[1], new_record.get_id());
+        assert_eq!(container.order[2], records[2].get_id());
     }
 }
