+++ relm4-store-collections/src/data_container/tests/test_data.rs
@@ -2,7 +2,7 @@ use backend_dummy::test_cases::TestRecord;
 use record::Record;
 
 use crate::WindowChangeset;
-use crate::implementation::data_container::DataContainer;
+use crate::data_container::DataContainer;
 
 
 #[cfg(test)]
